requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
# awscli
flake8
python-dotenv>=0.5.1
dvc
dvc-gdrive

# api requirements
fastapi
uvicorn[standard]

# others
numpy
matplotlib
Pillow
pickle
tensorflow
